Deploy doorstep render

+++ biohub/app/Http/Controllers/ServiceController.php
@@ -4,9 +4,12 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Service;
+use App\Http\Requests\StoreServiceRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ServiceController extends Controller
 {
+     use AuthorizesRequests;
     // Show all services
     public function index()
     {
@@ -22,13 +25,8 @@ class ServiceController extends Controller
     }
 
     // Save new service
-    public function store(Request $request)
+    public function store(StoreServiceRequest $request)
     {
-        $request->validate([
-            'name'             => 'required|string|max:255',
-            'duration_minutes' => 'required|integer|min:15',
-            'price'            => 'required|numeric|min:0',
-        ]);
 
         $provider = auth()->user()->providerProfile;
 
@@ -45,15 +43,13 @@ class ServiceController extends Controller
 
     // Delete a service
     public function destroy($id)
-    {
-        $provider = auth()->user()->providerProfile;
-        $service = Service::where('id', $id)
-            ->where('provider_id', $provider->id)
-            ->firstOrFail();
+{
+    $service = Service::findOrFail($id);
+    $this->authorize('delete', $service);
 
-        $service->delete();
+    $service->delete();
 
-        return redirect()->route('provider.services.index')
-            ->with('success', 'Service deleted!');
-    }
+    return redirect()->route('provider.services.index')
+        ->with('success', 'Service deleted!');
+}
 }
